+++ src/main/java/io/jenkins/plugins/analysis/core/model/ResultAction.java
@@ -77,6 +77,15 @@ public class ResultAction implements HealthReportingAction, LastBuildAction, Run
     }
 
     /**
+     * Returns the name of the static analysis tool.
+     *
+     * @return the ID
+     */
+    public String getName() {
+        return name;
+    }
+
+    /**
      * Returns the associated build/run that created the static analysis result.
      *
      * @return the run
@@ -151,6 +160,7 @@ public class ResultAction implements HealthReportingAction, LastBuildAction, Run
      *
      * @return {@code true} if a large image is defined, {@code false} otherwise
      */
+    @SuppressWarnings("unused") // Called by jelly view
     public boolean hasLargeImage() {
         return StringUtils.isNotBlank(getLargeImageName());
     }
@@ -160,6 +170,7 @@ public class ResultAction implements HealthReportingAction, LastBuildAction, Run
      *
      * @return the URL of the image
      */
+    @SuppressWarnings({"unused", "WeakerAccess"}) // Called by jelly view
     public String getLargeImageName() {
         return getLabelProvider().getLargeIconUrl();
     }
@@ -169,6 +180,7 @@ public class ResultAction implements HealthReportingAction, LastBuildAction, Run
      *
      * @return the URL of the image
      */
+    @SuppressWarnings("unused") // Called by jelly view
     public String getSmallImageName() {
         return getSmallImage();
     }
@@ -178,6 +190,7 @@ public class ResultAction implements HealthReportingAction, LastBuildAction, Run
      *
      * @return the URL of the image
      */
+    @SuppressWarnings({"unused", "WeakerAccess"}) // Called by jelly view
     public String getSmallImage() {
         return getLabelProvider().getSmallIconUrl();
     }
@@ -207,7 +220,12 @@ public class ResultAction implements HealthReportingAction, LastBuildAction, Run
         return String.format("%s for %s", getClass().getName(), getLabelProvider().getName());
     }
 
-    private StaticAnalysisLabelProvider getLabelProvider() {
+    /**
+     * Returns the {@link StaticAnalysisLabelProvider} for this action.
+     *
+     * @return the label provider for this tool
+     */
+    public StaticAnalysisLabelProvider getLabelProvider() {
         return new LabelProviderFactory().create(id, name);
     }